backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
email-validator
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
pytest
pytest-asyncio
httpx
pytest-cov